EASI-PASS Batch 2: rename src/ package -> easipass/
