+++ easipass/auto_stitching.py
