+++ easipass/functional.py
