+++ easipass/hcr_centroid_registration.py
