+++ easipass/somaprint_hcr.py
